move serveNPM to its own module
